refactor(scripts): one export per file, rename Bump to ReleaseType
